+++ containers/db/initialize/restore.d/04-create-deduplication-db.sql
@@ -32,6 +32,23 @@ CREATE TABLE nbs_mpi_mapping (
 );
 GO
 
+CREATE TABLE merge_group (
+  id BIGINT IDENTITY(1,1) PRIMARY KEY,
+  add_time DATETIME DEFAULT GETDATE()
+);
+GO
+
+CREATE TABLE merge_group_entries (
+	id BIGINT IDENTITY(1, 1) PRIMARY KEY,
+	merge_group BIGINT NOT NULL,
+	person_uid BIGINT NOT NULL,
+	is_merge BIT NULL,
+	last_chg_time DATETIME NULL,
+	last_chg_user_id BIGINT NULL,
+	FOREIGN KEY (group_id) REFERENCES merge_group(id)
+);
+GO
+
 CREATE TABLE matches_requiring_review (
   id BIGINT IDENTITY(1,1) PRIMARY KEY,
   merge_group BIGINT NOT NULL,
@@ -42,15 +59,6 @@ CREATE TABLE matches_requiring_review (
   date_identified DATETIME DEFAULT GETDATE(),
   matched_person_uid BIGINT NOT NULL,
   FOREIGN KEY (merge_group) REFERENCES merge_group(id)
-);
-GO
-
-CREATE TABLE merge_group (
-  id BIGINT IDENTITY(1,1) PRIMARY KEY,
-  person_uid BIGINT NOT NULL,
-  is_merge BIT NULL,
-  last_chg_time DATETIME NULL,
-  last_chg_user_id BIGINT NULL
 );
 GO
