+++ deduplication/src/main/java/gov/cdc/nbs/deduplication/batch/model/MatchesRequireReviewResponse.java
@@ -4,12 +4,13 @@ import java.util.ArrayList;
 import java.util.List;
 
 public record MatchesRequireReviewResponse(List<MatchRequiringReview> matches, int page, int total) {
+
   public MatchesRequireReviewResponse(int page, int total) {
     this(new ArrayList<>(), page, total);
   }
 
   public record MatchRequiringReview(
-      Long matchId,
+      Long groupId,
       String patientId,
       String patientLocalId,
       String patientName,
