+++ deduplication/src/main/java/gov/cdc/nbs/deduplication/batch/service/PatientRecordService.java
@@ -1,23 +1,23 @@
 package gov.cdc.nbs.deduplication.batch.service;
 
-import gov.cdc.nbs.deduplication.auth.authentication.PermissionResolver;
-import gov.cdc.nbs.deduplication.batch.mapper.PersonMergeDataMapper;
-import gov.cdc.nbs.deduplication.batch.model.PersonMergeData;
-import gov.cdc.nbs.deduplication.constants.QueryConstants;
-import gov.cdc.nbs.deduplication.merge.model.PatientNameAndTime;
-import gov.cdc.nbs.deduplication.seed.mapper.MpiPersonMapper;
-import gov.cdc.nbs.deduplication.seed.model.MpiPerson;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
-import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
-import org.springframework.stereotype.Service;
-
 import java.sql.ResultSet;
 import java.sql.Timestamp;
 import java.time.LocalDateTime;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
+import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
+import org.springframework.stereotype.Service;
+
+import gov.cdc.nbs.deduplication.auth.authentication.PermissionResolver;
+import gov.cdc.nbs.deduplication.batch.mapper.PersonMergeDataMapper;
+import gov.cdc.nbs.deduplication.batch.model.PersonMergeData;
+import gov.cdc.nbs.deduplication.merge.model.PatientNameAndTime;
+import gov.cdc.nbs.deduplication.seed.mapper.MpiPersonMapper;
+import gov.cdc.nbs.deduplication.seed.model.MpiPerson;
 
 @Service
 public class PatientRecordService {
@@ -38,7 +38,7 @@ public class PatientRecordService {
         .addValue("personParentUid", personParentUid);
 
     List<MpiPerson> mpiPersons = namedParameterJdbcTemplate.query(
-        QueryConstants.PERSON_RECORD_BY_PARENT_ID,
+        PatientQueries.PERSON_RECORD_BY_PARENT_ID,
         params,
         mpiPersonMapper);
     return (mpiPersons.isEmpty()) ? null : mpiPersons.getFirst();// most recent based on last_chg_time
@@ -46,7 +46,7 @@ public class PatientRecordService {
 
   public MpiPerson fetchPersonRecord(String personUid) {
     List<MpiPerson> results = namedParameterJdbcTemplate.query(
-        QueryConstants.PERSON_RECORD_BY_PERSON_ID,
+        PatientQueries.PERSON_RECORD_BY_PERSON_ID,
         new MapSqlParameterSource("personUid", personUid),
         mpiPersonMapper);
     if (results.isEmpty()) {
@@ -60,7 +60,7 @@ public class PatientRecordService {
     MapSqlParameterSource params = new MapSqlParameterSource()
         .addValue("ids", personUids);
     return namedParameterJdbcTemplate.query(
-        QueryConstants.PERSON_RECORDS_BY_PERSON_IDS,
+        PatientQueries.PERSON_RECORDS_BY_PERSON_IDS,
         params,
         mpiPersonMapper);
   }
@@ -74,7 +74,7 @@ public class PatientRecordService {
         .addValue("programJurisdictionOids", programJurisdictionOids);
 
     return namedParameterJdbcTemplate.query(
-        QueryConstants.PERSONS_MERGE_DATA_BY_PERSON_IDS,
+        PatientQueries.PERSONS_MERGE_DATA_BY_PERSON_IDS,
         params,
         new PersonMergeDataMapper(hasHivAccess));
   }
@@ -84,7 +84,7 @@ public class PatientRecordService {
         .addValue("ids", personUids);
 
     List<Map<String, Object>> result = namedParameterJdbcTemplate.queryForList(
-        QueryConstants.FETCH_PATIENT_ADD_TIME_QUERY, params);
+        PatientQueries.FETCH_PATIENT_ADD_TIME_QUERY, params);
 
     Map<String, LocalDateTime> addTimeMap = new HashMap<>();
     for (Map<String, Object> row : result) {
@@ -97,7 +97,7 @@ public class PatientRecordService {
 
   public PatientNameAndTime fetchPersonNameAndAddTime(String id) {
     return namedParameterJdbcTemplate.query(
-        QueryConstants.FIND_NBS_ADD_TIME_AND_NAME_QUERY,
+        PatientQueries.FIND_NBS_ADD_TIME_AND_NAME_QUERY,
         new MapSqlParameterSource()
             .addValue("id", id),
         (ResultSet rs, int rowNum) -> new PatientNameAndTime(
