+++ deduplication/src/test/java/gov/cdc/nbs/deduplication/merge/MatchesRequiringReviewResolverTest.java
@@ -71,7 +71,7 @@ class MatchesRequiringReviewResolverTest {
     assertThat(params.getValue("limit")).isEqualTo(13);
     assertThat(params.getValue("offset")).isEqualTo(13);
     assertThat(response.matches()).hasSize(1);
-    assertThat(response.matches().get(0).matchId()).isEqualTo(1l);
+    assertThat(response.matches().get(0).groupId()).isEqualTo(1l);
     assertThat(response.matches().get(0).patientId()).isEqualTo("12345");
     assertThat(response.matches().get(0).patientLocalId()).isEqualTo("444");
     assertThat(response.matches().get(0).patientName()).isEqualTo("John Doe");
